admin: move the `serverApp` file to typescript
This file configures the express application and is now correctly typed
in typescript.
