+++ packages/evolution-backend/package.json
@@ -39,6 +39,7 @@
         "moment": "^2.30.1",
         "morgan": "^1.10.0",
         "papaparse": "^5.4.1",
+        "passport": "^0.7.0",
         "request-ip": "^3.3.0",
         "serve-favicon": "^2.5.0",
         "slugify": "^1.6.6",
